requirements.txt
aiogram>=3.3.0,<4.0
anthropic>=0.40.0
tavily-python>=0.5.0
APScheduler>=3.10.0,<4.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
structlog>=24.0.0
